pyproject.toml
[project]
name = "fast-movie-reserve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "celery>=5.5.3",
    "faker>=37.8.0",
    "fastapi[standard]>=0.118.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "resend>=2.15.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
